Split code into public and private parts
